TAC-Bot-V2/requirements.txt
discord.py==0.16.12
PyNaCl==1.3.0
python-utils==2.3.0

